feat: add health check test
