+++ Cargo.toml
@@ -20,3 +20,10 @@ serde_with = { version = "1.4.0", features = ["json"] }
 derive_more = "0.99.17"
 futures = "0.3.15"
 hyper = "0.14.9"
+
+[lib]
+path = "src/lib.rs"
+
+[[bin]]
+path = "src/main.rs"
+name = "realioverse_api"
